feat: add post-confirmation dispute payload builder
